Split up topics actionTypes into different files.
