+++ app/modules/topics/actionTypes/index.js
@@ -1,0 +1,11 @@
+// @flow
+
+import { type ReducerAction } from './reducer';
+import { type ApiSagaAction } from './apiSaga';
+import { type TaskSagaAction } from './taskSaga';
+
+export type ModuleAction = ReducerAction | ApiSagaAction | TaskSagaAction;
+
+export * from './reducer';
+export * from './apiSaga';
+export * from './taskSaga';
